-Id ya no es un int es un Long -Hemos añadido el delete y optimizado el codigo del resto del CRUD

+++ src/main/java/org/iesch/Productos/repositorio/ProductoRepositorio.java
@@ -3,5 +3,5 @@ package org.iesch.Productos.repositorio;
 import org.iesch.Productos.modelo.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductoRepositorio extends JpaRepository<Producto, Integer> {
+public interface ProductoRepositorio extends JpaRepository<Producto, Long> {
 }
